# admin-panel/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Shortlinker API",
    "description": "Shortlinker management API",
    "contact": {
      "name": "AptS-1547",
      "email": "apts-1547@esaps.net"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.5.2"
  },
  "paths": {
...
    "/admin/v1/links": {
      "get": {
        "tags": [
          "links"
        ],
        "summary": "获取所有链接（支持分页和过滤）",
        "operationId": "list_links",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "minimum": 0
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "minimum": 0
            }
          },
          {
            "name": "created_after",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "created_before",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "only_expired",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "boolean",
                "null"
              ]
            }
          },
          {
            "name": "only_active",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "boolean",
                "null"
              ]
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated short links",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_Vec_LinkResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid filter"
          }
        }
      },
      "post": {
        "tags": [
          "links"
        ],
        "summary": "创建新链接",
        "operationId": "create_link",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostNewLink"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Short link created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_PostNewLink"
                }
              }
            }
          },
          "400": {
            "description": "Invalid short link"
          },
          "409": {
            "description": "Short code already exists"
          }
        }
      }
    },
...
    "/admin/v1/links/{code}": {
      "get": {
        "tags": [
          "links"
        ],
        "summary": "获取单个链接",
        "operationId": "get_link",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "description": "Short code",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Short link",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_LinkResponse"
                }
              }
            }
          },
          "404": {
            "description": "Short link not found"
          }
        }
      },
      "put": {
        "tags": [
          "links"
        ],
        "summary": "更新链接",
        "operationId": "update_link",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "description": "Short code",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostNewLink"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Short link updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_PostNewLink"
                }
              }
            }
          },
          "404": {
            "description": "Short link not found"
          }
        }
      },
      "delete": {
        "tags": [
          "links"
        ],
        "summary": "删除链接",
        "operationId": "delete_link",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "description": "Short code",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Short link deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_MessageResponse"
                }
              }
            }
          },
          "404": {
            "description": "Short link not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse_LinkResponse": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "type": "object",
            "required": [
              "code",
              "target",
              "created_at",
              "expires_at",
              "password",
              "click_count"
            ],
            "properties": {
              "click_count": {
                "type": "integer",
                "minimum": 0
              },
              "code": {
                "type": "string"
              },
              "created_at": {
                "type": "string"
              },
              "expires_at": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "password": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "target": {
                "type": "string"
              }
            }
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ApiResponse_MessageResponse": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "type": "object",
            "description": "简单消息响应",
            "required": [
              "message"
            ],
            "properties": {
              "message": {
                "type": "string"
              }
            }
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ApiResponse_PostNewLink": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "type": "object",
            "required": [
              "target"
            ],
            "properties": {
              "code": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "expires_at": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "force": {
                "type": [
                  "boolean",
                  "null"
                ]
              },
              "password": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "target": {
                "type": "string"
              }
            }
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "PaginatedResponse_Vec_LinkResponse": {
        "type": "object",
        "required": [
          "code",
          "message",
          "pagination"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "code",
                "target",
                "created_at",
                "expires_at",
                "password",
                "click_count"
              ],
              "properties": {
                "click_count": {
                  "type": "integer",
                  "minimum": 0
                },
                "code": {
                  "type": "string"
                },
                "created_at": {
                  "type": "string"
                },
                "expires_at": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "password": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "target": {
                  "type": "string"
                }
              }
            }
          },
          "message": {
            "type": "string"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationInfo"
          }
        }
      },
      "PaginationInfo": {
        "type": "object",
        "required": [
          "page",
          "page_size",
          "total",
          "total_pages"
        ],
        "properties": {
          "page": {
            "type": "integer",
            "minimum": 0
          },
          "page_size": {
            "type": "integer",
            "minimum": 0
          },
          "total": {
            "type": "integer",
            "minimum": 0
          },
          "total_pages": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "PostNewLink": {
        "type": "object",
        "required": [
          "target"
        ],
        "properties": {
          "code": {
            "type": [
              "string",
              "null"
            ]
          },
          "expires_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "force": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "password": {
            "type": [
              "string",
              "null"
            ]
          },
          "target": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}